Changes before combine InventoryBase and InventoryBaseUI
