--- Assets/Scripts/UI/Inventory.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 28de2bfa856663e4f8d3f38f080ba820
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
